mm: writeback: Disable dirty_writback when screen is on
every 500ms the writeback daemon kicks in and flushes the dirty pages into disk.
Thats not very pleasant for performance and *maybe* battery. This patch disables
the writeback function while the screen is on, and enables it back when the screen
is off to flush all the dirty pages. In my real world usage testing I notice a
slight bump in smoothness and performance while using the device.

Change-Id: I9233a6bc7417853343e87fcf0fc3c61e271e8869
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>